Tweaks to the Order Model
Tweaks to the Console Commands

+++ routes/console.php
@@ -39,7 +39,7 @@ Artisan::command('cleanVehicle', function () {
         }
         $bar->finish();
     } else {
-        $this->line('No Vehicles require cleaning');
+        $this->info('No Vehicles in recycle bin require cleaning');
     }
 })->purpose(
     'Remove Vehicles if they are still in the recycle bin after a month',
@@ -51,22 +51,67 @@ Artisan::command('cleanOrders', function () {
         ->onlyTrashed()
         ->get();
     if (count($deleted) !== 0) {
-        $this->line('Clearing orders deleted more than a month ago...');
-        $bar = $this->output->createProgressBar(count($deleted));
-        $bar->start();
-        foreach ($deleted as $item) {
-            $item->forceDelete();
-            $bar->advance();
+        if (
+            $this->confirm(
+                $deleted->count() .
+                    ' orders have been in the recycle bin for more than a month. Would you like to remove them?',
+                true,
+            )
+        ) {
+            $this->line('Clearing orders deleted more than a month ago...');
+            $bar = $this->output->createProgressBar(count($deleted));
+            $bar->start();
+            foreach ($deleted as $item) {
+                $item->forceDelete();
+                $bar->advance();
+            }
+            $bar->finish();
         }
-        $bar->finish();
     } else {
-        $this->line('No Orders require cleaning');
+        $this->info('No Orders in the recycle bin require cleaning');
     }
 })->purpose('Remove orders if they are still in the recycle bin after a month');
 
+Artisan::command('noOrderCustomers', function () {
+    $date = Carbon::now()
+        ->subMonths(3)
+        ->toDateTimeString();
+
+    $customers = Customer::where('created_at', '<=', $date)
+        ->withCount('orders')
+        ->get();
+
+    $customers = $customers->where('orders_count', 0);
+
+    if (count($customers) !== 0) {
+        if (
+            $this->confirm(
+                $customers->count() .
+                    ' customers have made no orders and been in the system for 3 months or more. Would you like to remove them?',
+                true,
+            )
+        ) {
+            $this->line('Clearing customers with no orders');
+            $bar = $this->output->createProgressBar(count($customers));
+            $bar->start();
+            foreach ($customers as $customer) {
+                $customer->delete();
+                $bar->advance();
+            }
+            $bar->finish();
+        }
+    } else {
+        $this->info(
+            'All customers on the system for more than three months have orders',
+        );
+    }
+});
+
 Artisan::command('cleanCustomers', function () {
-    $date = Carbon::now()->subMonth();
-    $deleted = Customer::where('deleted_at', '<', $date)
+    $date = Carbon::now()
+        ->subMonth()
+        ->toDateTimeString();
+    $deleted = Customer::where('deleted_at', '<=', $date)
         ->onlyTrashed()
         ->get();
     if (count($deleted) !== 0) {
@@ -79,7 +124,7 @@ Artisan::command('cleanCustomers', function () {
         }
         $bar->finish();
     } else {
-        $this->line('No Customers require cleaning');
+        $this->info('No Customers in recycle bin require cleaning');
     }
 })->purpose('Remove orders if they are still in the recycle bin after a month');
 
@@ -87,6 +132,7 @@ Artisan::command('databaseClean', function () {
     $this->call('cleanOrders');
     $this->call('cleanVehicle');
     $this->call('cleanCustomers');
+    $this->call('noOrderCustomers');
 });
 
 Artisan::command('checkReservationExpiry', function () {
